Using OrderedDict to preserve the order of fields
Airtable's API returns table records with the fields in a consistent
order. Notably, the tables' Name column always appears first among the
fields. But when the JSON response is deserialized into a regular
dictionary, the order of the fields is lost. If we want to preserve this
order, we can use the OrderedDict class with the object_pairs_hook
option when we load the JSON data.