Add MIT License
On https://pypi.org/project/airtable/ this project is listed as having the MIT license but this is not apparent from the GitHub repo. If the project is intended to be MIT licensed could you accept this PR? This would make it easier for us to get permission to use this project in our company and make it clearer to others viewing this repo what the intended license is.