Update README.rst (#61)
api keys are deprecated and will no longer be supported next year but using an access key as generated at https://airtable.com/developers/web/guides/personal-access-tokens in the airtable.Airtable method works just the same as the API_KEY currently does. lets shift the language to what it will be in the future.